refactor(types): split value responsibilities
